Completed lstm training. Testing mostly implemented but appears to have a small bug. Added some comments to LR code and edited helper_scripts slightly.

+++ LSTM/data_preprocessing.py
@@ -78,7 +78,7 @@ def create_batches(data, labels, interpolation, batch_size):
             batched_lengths.append(data_lengths[i*batch_size:])
 
     for i in range(len(batched_data)):
-        batched_data[i] = th.as_tensor(np.vstack(batched_data[i]))
+        batched_data[i] = th.as_tensor(np.vstack(batched_data[i])).float()
         batched_labels[i] = th.as_tensor(np.vstack(to_categorical(batched_labels[i],2)))
 
     return batched_data, batched_labels, batched_lengths
